Replace run shell script with logging scripts
The run script didn't do anything other than to document that you could
provide an explicit MAC address. A more useful script (which we can still
include the MAC note in) is one which logs output to a file. Also include
a script to submit the most recent measurement to the GMC.MAP service.