Add validation of packet data after unpacking
Since we don't know the protocol with certianty, there is a chance
of things going off the rails and bogus data being handled. To
try and prevent this, perform some basic packet validation and
raise a ValueError if things don't seem right.

We add checks to make sure that the decoded data itself is within
"reasonable" bounds, as well as enforce the packet ID to only
increment by 1. If an error occurs something has gone wrong and
we need to raise an exception so somebody can do something about
it.