Add synchronization code to recover from packet decode errors
If for some reason we can't decode a packet, its possible that we've
become desynced somehow. Try to recover from this situation by "nibbling"
our way through the buffer until we are able to make a few sucessfull
decodes in a row.

Also sync at the beginning of a new connection since we don't necessarily
know if the data we get from the device is going to start at a packet
boundary or not. While the connection starts out synchronized, I've noticed
that after a while it can drift. This isn't a problem during steady-state
since we consume whole packets at a time and buffer any remaining bytes
for the next round, but after a connection interruption we can't guarantee
that the data is packet-aligned.