Rename "util" package to "_util" so that its private nature is explicit
The "util" package is not meant for outside code to rely on. It is for
internal use by the library and its examples only. The interfaces in
this package can change at any time, so to make this clear to outside
users we prepend an underscore to its name.

Outside implementations are of course free to copy and paste the code
in this package into their own projects, but it is not safe to rely
on it as a library.