Apply Python static type system to codebase and fix bugs
Decorates functions and variables with static type information to let
software like mypy perform additional sanity checks. In the process of
adding this information and running mypy a few issues were found and
addressed.

Note that bluepy does not come with its own type definitions, so we've
had to create our own using the `stubgen` tool. The stub type code was
edited by hand to match the documentation, though it seems the docs
may be out of date (e.g. the list of functions don't match up with the
code...)

You can test the static code validator by running:

    $ mypy radalert ref