Make unknown field values non-fatal
Since we don't know what the unknown fields contain, it does not make
sense to treat unexpected behavior from them as fatal. We still want to
log when these events happen so that we can dig deeper into what they
might mean, however.

This commit replaces the ValueError exceptions on the unknown fields
with a loop that checks wether any unknown fields have taken on an
unexpected values. If it happens, we print out the entire buffer for
later analysis.