Improve exception handling of log example
Add a check for the BrokenPipeError which can be triggered by bluepy
when the connection is flaky. This shouldn't be fatal, so allow a
disconnect/reconnect cycle.

Also remove the example's spin method (which is basically useless
since the device spin does everything) so that we can succinctly
have a "finally" case which disconnects and sleeps for any exception
that we may encounter.