Increase bluetooth timeout from 4 seconds to 8.5 seconds
I've been having a lot of trouble with loosing connection and then not
being able to reconnect with the current 4 second timeout. Increasing
this to 8.5 seconds makes for a more reliable connection.

Note that the device seems to power down its Bluetooth radio after about
10 seconds of no communication, so there is some worry that if we loose
contact for 8.5 seconds and then sleep for 3 seconds (in the example)
we will expired the timer. Its hard to say but the larger timeout does
seem to be working much better regardless of these concerns.