Abstract BLE interface behind a "generic" abstract base class
The first step toward introducing support for the device's USB HID
interface is to define an abstract base class that will serve as the
parent for both HID and BLE classes. This will (somewhat) relieve
programs from having to care about which interface they are dealing
with.

Note that there are a number of features which are only present on
one interface or the other. Because of this, several functions are
forced to raise a `NotImplementedError` :(