Add HID backend for USB communication
This commit allows for you to connect to the geiger counter over a USB
HID connection. This supports different features than BTLE and also has
the benefit of providing a power source. Some parts of the protocol are
still unknown but the code / example provide most of the basic data.